backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.3
motor==3.4.0

# AI and ML
google-generativeai>=0.3.0
numpy>=1.24.3

# MCP (Model Context Protocol) for medical chatbot
mcp==1.16.0

# Background tasks
celery[mongodb]==5.3.4
kombu==5.3.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Security
cryptography>=42.0.0
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
